+++ test/pool.js
@@ -3,7 +3,30 @@
 const metatests = require('metatests');
 const metautil = require('..');
 
-metatests.test('Pool: add/next', async (test) => {
+metatests.test('Pool: add/next factor', async (test) => {
+  const pool = new metautil.Pool();
+
+  const obj1 = { a: 1 };
+  pool.add(obj1);
+  pool.setFactor(obj1, 0.5);
+  const obj2 = { a: 2 };
+  pool.add(obj2);
+  pool.setFactor(obj2, 0.2);
+  const obj3 = { a: 3 };
+  pool.add(obj3);
+  pool.setFactor(obj3, 0.7);
+
+  test.strictSame(await pool.next(), obj2);
+  pool.setFactor(obj2, 0.9);
+  test.strictSame(await pool.next(), obj1);
+  pool.setFactor(obj3, 0.2);
+  test.strictSame(await pool.next(), obj3);
+  pool.setFactor(obj2, 0.1);
+  test.strictSame(await pool.next(), obj2);
+  test.end();
+});
+
+metatests.test('Pool: add/next robin', async (test) => {
   const pool = new metautil.Pool();
 
   const obj1 = { a: 1 };
@@ -16,7 +39,35 @@ metatests.test('Pool: add/next', async (test) => {
   test.strictSame(await pool.next(), obj1);
   test.strictSame(await pool.next(), obj2);
   test.strictSame(await pool.next(), obj3);
+  test.end();
+});
+
+metatests.test('Pool: add/next mix', async (test) => {
+  const pool = new metautil.Pool();
+
+  const obj1 = { a: 1 };
+  pool.add(obj1, 0.2);
+  const obj2 = { a: 2 };
+  pool.add(obj2);
+  pool.setFactor(obj2, 0.5);
+  const obj3 = { a: 3 };
+  pool.add(obj3);
+
+  test.strictSame(await pool.next(), obj3);
+  pool.setFactor(obj1, 0);
   test.strictSame(await pool.next(), obj1);
+  test.strictSame(await pool.next(), obj3);
+  test.strictSame(await pool.next(), obj1);
+  pool.setFactor(obj1, 0.6);
+  test.strictSame(await pool.next(), obj3);
+  pool.setFactor(obj3, 0.9);
+  test.strictSame(await pool.next(), obj2);
+  pool.setFactor(obj1, 0);
+  test.strictSame(await pool.next(), obj1);
+  pool.setFactor(obj3, 0);
+  test.strictSame(await pool.next(), obj3);
+  test.strictSame(await pool.next(), obj1);
+  test.strictSame(await pool.next(), obj3);
   test.end();
 });